refactored fmp provider directory
