+++ openbb_sdk/providers/fmp/openbb_fmp/models/forex_eod.py
