+++ openbb_sdk/providers/fmp/openbb_fmp/models/risk_premium.py
